requirements.txt
asgiref==3.8.1
boto3==1.35.17
botocore==1.35.17
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
deepl==1.18.0
defusedxml==0.8.0rc2
dj-rest-auth==6.0.0
Django==5.1.1
django-allauth==64.2.1
django-filter==24.3
django-rest-auth==0.9.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
google-api-core==2.19.2
google-auth==2.34.0
google-cloud-core==2.4.1
google-cloud-translate==3.16.0
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.66.1
grpcio-status==1.66.1
idna==3.8
inflection==0.5.1
jmespath==1.0.1
jwcrypto==1.5.6
packaging==24.1
pillow==10.4.0
polib==1.2.0
proto-plus==1.24.0
protobuf==5.28.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.2
qrcode==7.4.2
requests==2.32.3
rsa==4.9
s3transfer==0.10.2
setuptools==74.1.1
six==1.16.0
social-auth-app-django==5.4.2
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
